# src/main/resources/swagger/swagger.yml
...
openapi: 3.0.0
info:
  title: Community REST API
  description: Community REST API
  version: "1.0"
...
servers:
  - url: http://localhost:8080/
  - url: https://localhost:8080/
################################################################################
#                               Components                                     #
################################################################################
components:
...
  parameters:
    OffsetParam:
      name: offset
      in: query
      description: Offset
      required: true
      schema:
        type: integer
        format: int32
        default: -1
    LimitParam:
      name: limit
      in: query
      description: Negative number means unlimited
      required: true
      schema:
        type: integer
        format: int32
        default: -1
...
  schemas:
...
    UserStatus:
      type: string
      enum:
        - active
        - inactive
      default: active
...
    User:
      type: object
      properties:
        id:
          type: string
          default: ""
        name:
          type: string
          default: ""
        mail:
          type: string
          default: ""
        lock:
          type: boolean
          default: false
        roleId:
          type: integer
          format: int32
          default: 0
        totalCount:
          type: integer
          format: int32
          default: 0
        status:
          $ref: "#/components/schemas/UserStatus"
    UserSession:
      type: object
      properties:
        Common.UUID:
          type: string
          default: ""
        Common.Name:
          type: string
          default: ""
        Common.Image:
          type: string
          default: ""
        Profile.Mail:
          type: string
          default: ""
        Profile.SAMAccount:
          type: string
          default: ""
        Profile.Phone:
          type: string
          default: ""
        Profile.Office:
          type: string
          default: ""
        Profile.DeptName:
          type: string
          default: ""
        Profile.Ename:
          type: string
          default: ""
        Profile.Cname:
          type: string
          default: ""
        status:
          type: string
          default: active
        group:
          type: array
          items:
            type: string
        isSystemAdmin:
          type: boolean
        systemAdminExpiredTime:
          type: integer
          format: int64
        medal:
          type: string
          default: ""
        certificates:
          type: array
          items:
            type: string
        title:
          type: string
          default: ""
        medalId:
          format: int32
          default: 0
        employeeType:
          type: string
          default: ""
...
    AutoCompleteUser:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        department:
          type: string
          default: ""
        ext:
          type: string
          default: ""
        mail:
          type: string
          default: ""
        imgAvatar:
          type: string
          default: ""
        status:
          $ref: "#/components/schemas/UserStatus"
...
    Tag:
      type: object
      properties:
        label:
          type: string
...
    PrivilegedCommunityForum:
      type: object
      required:
        - userId
        - communityId
        - forumId
        - forumName
      properties:
        userId:
          type: string
        communityId:
          type: integer
        forumId:
          type: integer
        forumName:
          type: string
...
    OutlookQueryData:
      type: object
      properties:
        query:
          type: string
        communityId:
          type: integer
          format: int32
          default: 0
        forumId:
          type: integer
          format: int32
          default: 0
    OutlookQueryResult:
      type: object
      properties:
        match:
          type: array
          items:
            $ref: "#/components/schemas/User"
        unMatch:
          $ref: "#/components/schemas/OutlookQueryUnmatchedResult"
    OutlookQueryUnmatchedResult:
      type: object
      properties:
        delta:
          type: array
          items:
            $ref: "#/components/schemas/User"
        others:
          type: array
          items:
            type: string
...
    UserAutocompleteQuery:
      type: object
      required:
        - q
        - limit
        - exclude
      properties:
        q:
          type: string
        limit:
          type: integer
          format: int32
        withAvatar:
          type: boolean
          default: false
        exclude:
          type: array
          items:
            type: string
...
    UserQueryDto:
      type: object
      properties:
        uids:
          type: array
          items:
            type: string
        samAccounts:
          type: array
          items:
            type: string
        names:
          type: array
          items:
            type: string
        withAvatar:
          type: boolean
          default: false
...
paths:
...
  /users/from-outlook:
    post:
      tags:
        - User
      summary: Search for the users from outlook by email
      operationId: searchOutlookUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutlookQueryData"
        description: Outlook query data
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutlookQueryResult"
        400:
          description: Invalid parameters
  /users/to-outlook:
    post:
      tags:
        - User
      summary: transfer to outlook recipient format
      operationId: transferToOutlookUser
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        description: uid list
        required: true
      responses:
        200:
          description: OK
          content:
            "*/*":
              schema:
                type: string
        400:
          description: Invalid parameters
  /users:
    get:
      tags:
        - User
      summary: Search for the users by their name
      operationId: searchUserByName
      parameters:
        - name: q
          in: query
          description: query name
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/LimitParam"
        - name: exclude
          in: query
          description: Exclusion
          required: false
          schema:
            type: array
            items:
              type: string
        - name: withAvatar
          in: query
          description: Whether there is avatar in the response
          required: true
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutoCompleteUser"
        400:
          description: Invalid parameters
    post:
      tags:
        - User
      summary: Search for the users by their name
      operationId: getAutocompleteUserList
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAutocompleteQuery"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutoCompleteUser"
        400:
          description: Invalid parameters
  /users/info:
    post:
      tags:
        - User
      summary: Search for the users by uid or samAccount
      operationId: getUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserQueryDto"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSession"
        400:
          description: Invalid parameters
...
  /tag:
    get:
      tags:
        - Tag
      summary: get tags
      operationId: getTags
      parameters:
        - name: q
          in: query
          description: query name
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/LimitParam"
        - name: exclude
          in: query
          description: Exclusion
          required: true
          schema:
            type: array
            items:
              type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
        400:
          description: Invalid parameters
...
  /privilegedCommunityForum/getPrivilegedCommunityForumIds:
    get:
      tags:
        - PrivilegedCommunityForum
      summary: Get the list of privileged community and forum IDs
      description: Get the list of privileged community and forum IDs (called by mes)
      operationId: getPrivilegedCommunityForumIds
      parameters:
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrivilegedCommunityForum"
        204:
          description: No Content
        300:
          description: Multiple Choices
...
  /personwordcloud:
    get:
      tags:
        - Personwordcloud
      summary: get the data of personWordCloud
      operationId: getPersonWordCloud
      parameters:
        - name: id
          in: query
          description: userId
          required: true
          schema:
            type: string
        - name: topn
          in: query
          description: topn
          required: true
          schema:
            type: integer
            format: int32
        - name: lang
          in: query
          description: system language
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
...
  /integrate/checkCommunityHasAppGroup:
    get:
      tags:
        - Integration
      summary: =For 3/2 Integration only, Please contact Yvonne.Zheng to run= Create 2 UserGroup for each community
      operationId: createCommunityAppGroup
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
        400:
          description: Invalid parameters
  /integrate/community:
    get:
      tags:
        - Integration
      summary: =For 3/2 Integration only, Please contact Yvonne.Zheng to run= Integrate for community uid to gid
      operationId: integrateCommunityRole
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
        400:
          description: Invalid parameters
  /integrate/checkForumHasAppGroup:
    get:
      tags:
        - Integration
      summary: =For 3/2 Integration only, Please contact Yvonne.Zheng to run= Create 2 UserGroup for each forum
      operationId: createForumAppGroup
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
        400:
          description: Invalid parameters
  /integrate/forum:
    get:
      tags:
        - Integration
      summary: =For 3/2 Integration only, Please contact Yvonne.Zheng to run= Integrate for forum uid to gid
      operationId: integrateForumRole
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
        400:
          description: Invalid parameters
  /integrate/ddf:
    get:
      tags:
        - Integration
      summary: =For Integration only, Please contact Yvonne.Zheng to run= Integrate for ddf uid to gid
      operationId: integrateDdf
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
        400:
          description: Invalid parameters
  /integrate/orgCommunity:
    get:
      tags:
        - Integration
      summary: =For Integration only, Please contact Yvonne.Zheng to run= Integrate for ddf uid to gid
      operationId: integrateOrgCommunity
      parameters:
        - name: startTime
          in: query
          required: true
          schema:
            type: integer
            format: int64
            default: -1
        - name: endTime
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: integer
                format: int32
        400:
          description: Invalid parameters
...